refactor: reverser dependency for project and import-export
