+++ server/translations/src/import-export/dto/export-translations.dto.ts
@@ -1,0 +1,3 @@
+export interface ExportTranslationsDto {
+  projectId: number;
+}
