+++ server/translations/src/import-export/dto/import-translations.dto.ts
@@ -1,0 +1,3 @@
+export interface ImportTranslationsDto {
+  projectId: number;
+}
