+++ server/translations/src/projects/projects.controller.ts
@@ -7,11 +7,9 @@ import {
   Param,
   Post,
   Put,
-  Query,
 } from "@nestjs/common";
 import { ApiTags } from "@nestjs/swagger";
 import { CreateProjectDto } from "./dto/create-project.dto";
-import { ImportFileDto } from "./dto/import-file.dto";
 import { UpdateProjectDto } from "./dto/update-project.dto";
 import { ProjectsService } from "./projects.service";
 
@@ -51,23 +49,5 @@ export class ProjectsController {
   @Delete(":id")
   remove(@Param("id") id: string) {
     return this.projectsService.remove(+id);
-  }
-
-  @Post(":id/import")
-  importFile(@Param("id") id: string, @Body() importFileDto: ImportFileDto) {
-    logger.log(`POST /projects/${id}/import ${JSON.stringify(importFileDto)}`);
-    return this.projectsService.importFileToProject(+id, importFileDto);
-  }
-
-  @Get(":id/export")
-  exportTranslations(@Param("id") id: string, @Query("lang") lang: string) {
-    logger.log(`GET /projects/${id}/export?lang=${lang}`);
-    return this.projectsService.exportTranslations(+id, lang);
-  }
-
-  @Post(":id/github-import")
-  importGithubToProject(@Param("id") id: string) {
-    logger.log(`POST /projects/${id}/github-import`);
-    return this.projectsService.importGithubToProject(+id);
   }
 }
