+++ server/translations/src/projects/projects.module.ts
@@ -1,26 +1,14 @@
 import { Module } from "@nestjs/common";
-import { ClientsModule, Transport } from "@nestjs/microservices";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { ImportExportModule } from "../import-export/import-export.module";
 import { TranslationsModule } from "../translations/translations.module";
 import { ProjectRepository } from "./project.repository";
 import { ProjectsController } from "./projects.controller";
 import { ProjectsService } from "./projects.service";
 
 @Module({
-  imports: [
-    TypeOrmModule.forFeature([ProjectRepository]),
-    ClientsModule.register([
-      {
-        name: "INTEGRATION_SERVICE",
-        transport: Transport.TCP,
-        options: { port: 8092 },
-      },
-    ]),
-    ImportExportModule,
-    TranslationsModule,
-  ],
+  imports: [TypeOrmModule.forFeature([ProjectRepository]), TranslationsModule],
   controllers: [ProjectsController],
   providers: [ProjectsService],
+  exports: [ProjectsService],
 })
 export class ProjectsModule {}
